refactor: Incoroprate general fixes from mapdata rewrite
This is a collection of changes I have made in the mapdata rewrite branch, that is not really related to mapdata. They are causing unnecessary merge conflicts, so it's better to get them into main.
