+++ common/src/main/java/com/wynntils/core/json/JsonManager.java
@@ -10,15 +10,9 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 import com.google.gson.JsonParser;
-import com.google.gson.TypeAdapter;
-import com.google.gson.TypeAdapterFactory;
-import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonToken;
-import com.google.gson.stream.JsonWriter;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Manager;
-import com.wynntils.core.persisted.config.NullableConfig;
 import com.wynntils.utils.EnumUtils;
 import com.wynntils.utils.FileUtils;
 import com.wynntils.utils.colors.CustomColor;
@@ -40,7 +34,7 @@ import org.apache.commons.lang3.RandomStringUtils;
 public final class JsonManager extends Manager {
     public static final Gson GSON = new GsonBuilder()
             .registerTypeAdapter(CustomColor.class, new CustomColor.CustomColorSerializer())
-            .registerTypeAdapterFactory(new EnumTypeAdapterFactory<>())
+            .registerTypeAdapterFactory(new EnumUtils.EnumTypeAdapterFactory<>())
             .enableComplexMapKeySerialization()
             .setPrettyPrinting()
             .serializeNulls()
@@ -128,70 +122,5 @@ public final class JsonManager extends Manager {
                         "invalid_" + LocalDateTime.now().format(DateTimeFormatter.ISO_LOCAL_DATE_TIME) + "_"
                                 + RandomStringUtils.randomAlphanumeric(5) + "_" + jsonFile.getName()));
         FileUtils.deleteFile(jsonFile);
-    }
-
-    public static final class EnumTypeAdapterFactory<E extends Enum<E>> implements TypeAdapterFactory {
-        @Override
-        public <T> TypeAdapter<T> create(Gson gson, TypeToken<T> type) {
-            if (!type.getRawType().isEnum()) return null;
-
-            Class<E> enumClazz = (Class<E>) type.getRawType();
-            return (TypeAdapter<T>) new EnumTypeAdapter<>(enumClazz);
-        }
-    }
-
-    private static final class EnumTypeAdapter<T extends Enum<T>> extends TypeAdapter<T> {
-        private final Class<T> enumClazz;
-        private final boolean nullAllowed;
-
-        private EnumTypeAdapter(Class<T> enumClazz) {
-            this.enumClazz = enumClazz;
-            nullAllowed = NullableConfig.class.isAssignableFrom(enumClazz);
-        }
-
-        @Override
-        public void write(JsonWriter out, T value) throws IOException {
-            if (value != null) {
-                out.value(EnumUtils.toJsonFormat((Enum<?>) value));
-            } else {
-                if (!nullAllowed) {
-                    WynntilsMod.warn("Writing null enum value to json for " + enumClazz.getSimpleName());
-                }
-                out.nullValue();
-            }
-        }
-
-        @Override
-        public T read(JsonReader in) throws IOException {
-            if (in.peek() == JsonToken.STRING) {
-                String jsonString = in.nextString();
-
-                T value = EnumUtils.fromJsonFormat(enumClazz, jsonString);
-                if (value == null) {
-                    WynntilsMod.warn("Illegal enum value: " + jsonString + " for type " + enumClazz.getName());
-                    return replacement();
-                }
-
-                return value;
-            } else if (in.peek() == JsonToken.NULL) {
-                in.nextNull();
-
-                if (!nullAllowed) {
-                    WynntilsMod.warn("Null enum value for type " + enumClazz.getName());
-                    return replacement();
-                }
-
-                return null;
-            } else {
-                WynntilsMod.warn("Invalid json type " + in.peek() + " for enum " + enumClazz.getName());
-                return replacement();
-            }
-        }
-
-        private T replacement() {
-            Enum<?> firstValue = enumClazz.getEnumConstants()[0];
-            WynntilsMod.warn("Will replace with first enum value: " + firstValue.name());
-            return (T) firstValue;
-        }
     }
 }
