+++ common/src/main/java/com/wynntils/utils/EnumUtils.java
@@ -5,6 +5,16 @@
 package com.wynntils.utils;
 
 import com.google.common.base.CaseFormat;
+import com.google.gson.Gson;
+import com.google.gson.TypeAdapter;
+import com.google.gson.TypeAdapterFactory;
+import com.google.gson.reflect.TypeToken;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+import com.wynntils.core.WynntilsMod;
+import com.wynntils.core.persisted.config.NullableConfig;
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 import org.apache.commons.lang3.text.WordUtils;
@@ -52,6 +62,71 @@ public final class EnumUtils {
             return true;
         } catch (NoSuchMethodException e) {
             return false;
+        }
+    }
+
+    public static final class EnumTypeAdapterFactory<E extends Enum<E>> implements TypeAdapterFactory {
+        @Override
+        public <T> TypeAdapter<T> create(Gson gson, TypeToken<T> type) {
+            if (!type.getRawType().isEnum()) return null;
+
+            Class<E> enumClazz = (Class<E>) type.getRawType();
+            return (TypeAdapter<T>) new EnumTypeAdapter<>(enumClazz);
+        }
+    }
+
+    private static final class EnumTypeAdapter<T extends Enum<T>> extends TypeAdapter<T> {
+        private final Class<T> enumClazz;
+        private final boolean nullAllowed;
+
+        private EnumTypeAdapter(Class<T> enumClazz) {
+            this.enumClazz = enumClazz;
+            nullAllowed = NullableConfig.class.isAssignableFrom(enumClazz);
+        }
+
+        @Override
+        public void write(JsonWriter out, T value) throws IOException {
+            if (value != null) {
+                out.value(EnumUtils.toJsonFormat((Enum<?>) value));
+            } else {
+                if (!nullAllowed) {
+                    WynntilsMod.warn("Writing null enum value to json for " + enumClazz.getSimpleName());
+                }
+                out.nullValue();
+            }
+        }
+
+        @Override
+        public T read(JsonReader in) throws IOException {
+            if (in.peek() == JsonToken.STRING) {
+                String jsonString = in.nextString();
+
+                T value = EnumUtils.fromJsonFormat(enumClazz, jsonString);
+                if (value == null) {
+                    WynntilsMod.warn("Illegal enum value: " + jsonString + " for type " + enumClazz.getName());
+                    return replacement();
+                }
+
+                return value;
+            } else if (in.peek() == JsonToken.NULL) {
+                in.nextNull();
+
+                if (!nullAllowed) {
+                    WynntilsMod.warn("Null enum value for type " + enumClazz.getName());
+                    return replacement();
+                }
+
+                return null;
+            } else {
+                WynntilsMod.warn("Invalid json type " + in.peek() + " for enum " + enumClazz.getName());
+                return replacement();
+            }
+        }
+
+        private T replacement() {
+            Enum<?> firstValue = enumClazz.getEnumConstants()[0];
+            WynntilsMod.warn("Will replace with first enum value: " + firstValue.name());
+            return (T) firstValue;
         }
     }
 }
